made `Phase.order` a PositiveSmallIntegerField

+++ simpl/games/models.py
@@ -150,7 +150,7 @@ class Phase(AbstractTimeStampedModel):
         'Game',
         related_name='phases'
     )
-    order = models.IntegerField(blank=True, null=True, db_index=True)
+    order = models.PositiveSmallIntegerField(blank=True, null=True, db_index=True)
 
     class Meta(object):
         ordering = ('game', 'order')
